issue
Commit Summary:

Title: Issue with AttributeError in love_encounter function

Body:
Encountered an AttributeError in the `love_encounter` function within the `abm.py` file. The error occurs when trying to add an edge between two entities in the network. The error message indicates that the `network_manager` attribute of the `human` object is `None`.

The issue is located in the following code block:

```python
for group_id in human.grp.keys():
    if group_id in entities:  # Check if the Group object exists
        # add edge between the two entities in the network
        human.network_manager.add_edge(human.id, other.id)
        group = entities[group_id]  # Get the Group object
        if other.id in group.members:
            break
```

The next steps would be to investigate why `human.network_manager` is `None` at this point in the code. This could involve checking the initialization of the `human` object and ensuring that `network_manager` is being properly set.